Extract Environmental, Suitability, and Realised Values at Sampling Locations

This function extracts values from the environmental state, target suitability, and realised target state at the locations defined by the sampling effort in a `SimulationObject`.

Usage

sim_extract(simulation_object, realised_extract_fun = mean)

Arguments

simulation_object

A `SimulationObject` containing the environmental state, target suitability, realised target state, and sampling effort.

fun

fun argument passed to extract for extracting realised values.

...

Additional arguments passed to extract for extracting realised values.

Value

A `SimulationObject` with updated `@effort` slot containing extracted values for environmental state, target suitability, and realised target state at each sampling location.

Details

This function extracts the values from the environmental state (`@state_env`), target suitability (`@state_target_suitability`), and realised target state (`@state_target_realised`) at the sampling locations defined in the `@effort` slot of the input `SimulationObject`. The extracted values are added as columns to the `@effort` slot.